% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-reports.R
\name{ps_export_style}
\alias{ps_export_style}
\title{Export open GIS style XML}
\usage{
ps_export_style(
  x,
  file,
  format = c("qml", "sld"),
  layer_type = c("head_raster", "depth_raster", "contours", "contour_support", "arrows",
    "wells", "support"),
  field = NULL,
  units = NULL,
  palette = NULL,
  breaks = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{x}{Object whose values inform default raster breaks.}

\item{file}{Output QML or SLD file.}

\item{format}{QGIS QML or standards-based SLD.}

\item{layer_type}{Styled layer type.}

\item{field}{Attribute used for labels or support categories.}

\item{units}{Optional label units.}

\item{palette}{Colors or color function.}

\item{breaks}{Optional explicit continuous breaks.}

\item{overwrite}{Permit replacement of an existing file.}
}
\value{
A \code{potentiomap_style_export} manifest. Optional properties can render
differently among GIS versions.
}
\description{
Export open GIS style XML
}
\examples{
r <- terra::rast(nrows = 2, ncols = 2, xmin = 0, xmax = 2,
                 ymin = 0, ymax = 2, crs = "EPSG:26920", vals = 1:4)
file <- tempfile(fileext = ".qml")
style <- ps_export_style(r, file, "qml", "head_raster", units = "m")
style$manifest
# GIS versions can render optional style properties differently.
}
