% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ps_sample_aoi}
\alias{ps_sample_aoi}
\title{Make the sample area-of-interest polygon}
\usage{
ps_sample_aoi()
}
\value{
A \code{terra::SpatVector} polygon in EPSG:26916.
}
\description{
Make the sample area-of-interest polygon
}
\examples{
aoi <- ps_sample_aoi()
aoi
}
