% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{ps_surfaces}
\alias{ps_surfaces}
\title{Extract interpolated surfaces}
\usage{
ps_surfaces(x)
}
\arguments{
\item{x}{A \code{potentiomap_result} or named list of \code{SpatRaster} surfaces.}
}
\value{
A named list of \code{terra::SpatRaster} objects.
}
\description{
Extract interpolated surfaces
}
\examples{
data("synthetic_wells")
pts <- ps_make_points(synthetic_wells, "x", "y", "gw_elevation",
                      "well_id", "EPSG:26916")
result <- ps_interpolate(pts, methods = "IDW", grid_res = 150,
                         return = "result")
ps_surfaces(result)
}
