% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{ps_arrow_vertices}
\alias{ps_arrow_vertices}
\title{Extract arrow base or tip points}
\usage{
ps_arrow_vertices(arrows, which = c("last", "first"), out_file = NULL)
}
\arguments{
\item{arrows}{A line \code{SpatVector} or path to a line vector file.}

\item{which}{\code{"first"} for arrow bases or \code{"last"} for arrow tips.}

\item{out_file}{Optional output vector path.}
}
\value{
A point \code{SpatVector}.
}
\description{
Extract arrow base or tip points
}
\examples{
data("synthetic_wells")
pts <- ps_make_points(synthetic_wells, "x", "y", "gw_elevation",
                      "well_id", "EPSG:26916")
s <- ps_interpolate(pts, methods = "IDW", grid_res = 100)
arrows <- ps_flow_arrows(s$IDW, res_factor = 4, scale = 60)
tips <- ps_arrow_vertices(arrows$arrows, which = "last")
tips
}
