% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{ps_contours}
\alias{ps_contours}
\title{Create contours from a surface raster}
\usage{
ps_contours(surface, interval = 1, levels = NULL)
}
\arguments{
\item{surface}{A \code{terra::SpatRaster} potentiometric surface.}

\item{interval}{Contour interval in map elevation units.}

\item{levels}{Optional explicit contour levels. When supplied, \code{interval} is
ignored.}
}
\value{
A line \code{terra::SpatVector} of contours.
}
\description{
Create contours from a surface raster
}
\examples{
data("synthetic_wells")
pts <- ps_make_points(synthetic_wells, "x", "y", "gw_elevation",
                      "well_id", "EPSG:26916")
s <- ps_interpolate(pts, methods = "IDW", grid_res = 100)
ctr <- ps_contours(s$IDW, interval = 1)
ctr
}
